+++ scripts/generate-ts.sh
@@ -5,6 +5,8 @@ OPENAPI_YAML_PATH=$APP_PATH/openapi
 OPENAPI_GENERATED_TS_PATH=$APP_PATH/src/types/generated
 
 SRC_FILES=(
+    "legacy/ojp-v1/ojp-fare-request.yml"
+
     "legacy/ojp-v1/ojp-lir-request.yml"
     "legacy/ojp-v1/ojp-lir-response.yml"
     
@@ -12,14 +14,16 @@ SRC_FILES=(
     "legacy/ojp-v1/ojp-ser-response.yml"
     
     "legacy/ojp-v1/ojp-shared.yml"
+    
     "legacy/ojp-v1/ojp-tir-response.yml"
-    "legacy/ojp-v1/ojp-trip.yml"
+    
+    "legacy/ojp-v1/ojp-tr-request.yml"
+    "legacy/ojp-v1/ojp-tr-response.yml"
 
     "ojp-shared.yml"
 
-    "ojp-fare-request.yml"
     "ojp-fare-response.yml"
-    
+
     "ojp-lir-request.yml"
     "ojp-lir-response.yml"